# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Hotel management system for users and admin.",
        "title": "Restful Hotel App API",
        "contact": {
            "name": "Kiet Asmara",
            "url": "http://www.swagger.io/support",
            "email": "kiet123pascal@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "paths": {
...
        "/bookings/:id": {
            "post": {
                "description": "Pay for a booking by ID using deposit or Xendit",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Pay Booking By ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Booking ID",
                        "name": "ID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "The input payment struct",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.PaymentReq"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.PaymentResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        },
        "/bookings/payments": {
            "get": {
                "description": "Get all user payments",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Booking"
                ],
                "summary": "Get all payments",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.PaymentsResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        },
...
        "/users/deposit": {
            "get": {
                "description": "Get all and refresh user deposit history",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get all Deposits",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.GetDepositResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Deposit to balance with Xendit invoice",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Deposit",
                "parameters": [
                    {
                        "type": "string",
                        "description": "JWT Token",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "The input deposit struct",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/service.DepositInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/handlers.DepositResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handlers.DepositResponse": {
            "type": "object",
            "properties": {
                "invoice": {
                    "$ref": "#/definitions/service.XenditResponse"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.ErrResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.GetDepositResponse": {
            "type": "object",
            "properties": {
                "deposit_amount": {
                    "type": "number"
                },
                "deposits": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Deposit"
                    }
                }
            }
        },
...
        "handlers.PaymentReq": {
            "type": "object",
            "properties": {
                "payment_method": {
                    "type": "string"
                }
            }
        },
        "handlers.PaymentResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "payment": {
                    "$ref": "#/definitions/model.Payment"
                }
            }
        },
        "handlers.PaymentsResponse": {
            "type": "object",
            "properties": {
                "payments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/model.Payment"
                    }
                }
            }
        },
...
        "model.Deposit": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "deposit_id": {
                    "type": "integer"
                },
                "invoice_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.Payment": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "booking_id": {
                    "type": "integer"
                },
                "invoice_id": {
                    "type": "string"
                },
                "payment_date": {
                    "type": "string"
                },
                "payment_id": {
                    "type": "integer"
                },
                "payment_method": {
                    "type": "string",
                    "enum": [
                        "deposit",
                        "xendit"
                    ]
                },
                "status": {
                    "type": "string"
                },
                "url": {
                    "type": "string"
                }
            }
        },
...
        "service.DepositInput": {
            "type": "object",
            "required": [
                "deposit_amount"
            ],
            "properties": {
                "deposit_amount": {
                    "type": "number"
                }
            }
        },
...
        "service.XenditResponse": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "customer": {
                    "type": "object",
                    "additionalProperties": {}
                },
                "expiry_date": {
                    "type": "string"
                },
                "external_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}